+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - Rename `Values` to `CalendarValues`
 - Prefix theme keys with `Calendar`
+- Use `onSelectDate` instead of `onSelectStartDate` and `onSelectEndDate`
 
 ## [0.1.1] - 2021-11-12
 ### Added